Main Script
Parses a pdf file showing shifts at work by name for the upcoming week/month/year. Sorts through the file and creates a Google Calendar event for each shift on the correct days, so that each days shifts can be viewed by smartphone through the calendar widget at a glance , instead of the tedious process of logging into a hotmial account and finding the correct email over and over.